requirements.txt
mitmproxy==10.1.6
pymongo==4.6.1
python-dotenv==1.0.0
